Move ResourceDictionary to its own code file
